requirements.txt
FlagEmbedding==1.2.9
FlagEmbedding==1.3.5
numpy==2.2.6
openai==1.82.1
pandas==2.2.3
protobuf==6.31.1
python-dotenv==1.1.0
rank_bm25==0.2.2
Requests==2.32.3
scikit_learn==1.2.1
torch==2.2.2
tqdm==4.65.0
transformers==4.48.1
